tarantoolctl connect doesn't report errors
This happens because console:eval swallows errors
and doesn't print them to stdout.

Also, console.connect() doesn't always call on_client_disconnect()
in case of errors, so the code that was meant to terminate local
session on errors, didn't do this in some cases and caused local
console to remain open.

Fixes #1586